host/requirements.txt
pyserial
spotipy
python-dotenv
